DROID-737: move probeID managing logic to ProbeIdManager
